requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr<2.0,>=1.3
SQLAlchemy<1.1.0,>=0.9.7
eventlet>=0.17.4
keystonemiddleware>=2.0.0
kombu>=3.0.7
Routes!=2.0,!=2.1,>=1.12.3;python_version=='2.7'
Routes!=2.0,>=1.12.3;python_version!='2.7'
WebOb>=1.2.3
PasteDeploy>=1.5.0
Paste
sqlalchemy-migrate>=0.9.6
netaddr>=0.7.12
netifaces>=0.10.4
httplib2>=0.7.5
lxml>=2.3
passlib
python-heatclient>=0.3.0
python-novaclient>=2.22.0
python-cinderclient>=1.2.2
python-keystoneclient>=1.6.0
python-swiftclient>=2.2.0
python-designateclient>=1.0.0
python-neutronclient<3,>=2.3.11
iso8601>=0.1.9
jsonschema!=2.5.0,<3.0.0,>=2.0.0
Jinja2>=2.6 # BSD License (3 clause)
pexpect!=3.3,>=3.1 # ISC License
oslo.config>=1.11.0 # Apache-2.0
oslo.context>=0.2.0 # Apache-2.0
oslo.i18n>=1.5.0 # Apache-2.0
oslo.serialization>=1.4.0 # Apache-2.0
oslo.service>=0.1.0 # Apache-2.0
oslo.utils>=1.9.0 # Apache-2.0
oslo.concurrency>=2.1.0 # Apache-2.0
MySQL-python;python_version=='2.7'
Babel>=1.3
six>=1.9.0
stevedore>=1.5.0 # Apache-2.0
oslo.messaging>=1.16.0 # Apache-2.0
osprofiler>=0.3.0 # Apache-2.0
oslo.log>=1.6.0 # Apache-2.0
